tests/test-requirements.txt
behave==1.2.5 \
    --hash=sha256:89238a5e4b11ff607e8ebc6b4b1fb1a0b1f3d794fba80e1fb4b6b3652979c927 \
    --hash=sha256:8c182feece4a519c5ffc11e1ab3682d25d5a390dd5f4573bb1296443beb9d7c7 \
    --hash=sha256:81b731ac5187e31e4aad2594944fa914943683a9818320846d037c5ebd6d5d0b \
    --hash=sha256:07c741f30497b6f9361a9bc74c68418507cd17e70d6f586faa3bff57684a2ec8
six==1.11.0 \
    --hash=sha256:832dc0e10feb1aa2c68dcc57dbb658f1c7e65b9b61af69048abc87a2db00a0eb \
    --hash=sha256:70e8a77beed4562e7f14fe23a786b54f6296e34344c23bc42f07b15018ff98e9
parse_type==0.4.2 \
    --hash=sha256:6e906a66f340252e4c324914a60d417d33a4bea01292ea9bbf68b4fc123be8c9 \
    --hash=sha256:f596bdc75d3dd93036fbfe3d04127da9f6df0c26c36e01e76da85adef4336b3c
parse==1.8.2 \
    --hash=sha256:8048dde3f5ca07ad7ac7350460952d83b63eaacecdac1b37f45fd74870d849d2
selenium==3.8.1 \
    --hash=sha256:5acb9cdbc2d1a7fbb3e16a8ce9246211cc371f0367ad9c6bc2273cca60a6b045 \
    --hash=sha256:9abd2dbd4a5e9b778483ce7e5adf1ea9364fcbc29da488e979213c825a1515d3
capybara-py==0.1.3 \
    --hash=sha256:c1c44b570e8494d0ce7b5d443e79828f2ade3150425190d044718916f11450d6
lxml==4.1.1 \
    --hash=sha256:6cba398eb37e0631e60e0e080c101cfe91769b2c8267105b64b4625e2581ea21 \
    --hash=sha256:e6b6698415c7e8d227a47a3b1038e1b37c2b438a1b48c2db7ad9e74ddbcd1149
xpath-py==0.1.2 \
    --hash=sha256:6e111a0be0df610d3fad7acf03f2e5dcaeec1b6d724a708a397f838dc0c463fa
cssselect==1.0.3 \
    --hash=sha256:e6b6698415c7e8d227a47a3b1038e1b37c2b438a1b48c2db7ad9e74ddbcd1149 \
    --hash=sha256:3b5103e8789da9e936a68d993b70df732d06b8bb9a337a05ed4eb52c17ef7206